# projects/neural-sync/artifacts/api-contracts.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEURAL SYNC API",
    "version": "1.0.0",
    "description": "Multi-dimensional AI-driven developer–project compatibility engine. All protected routes require JWT Bearer authentication per ADR-002. The 500 ms p95 latency SLA applies to the synchronous POST /matches response per ADR-003. LLM explanations are generated asynchronously (Phase-1: Google Gemini; Claude API is a supported alternate per ADR-004); the initial response always contains a deterministic stub (≥50 chars) per ADR-003/ADR-004. Error envelope on all 4xx/5xx: {error_code, message, request_id}.",
    "contact": {
      "name": "NEURAL SYNC Engineering",
      "email": "engineering@neural-sync.example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    },
    {
      "url": "https://neural-sync.example.com/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
...
  "paths": {
...
    "/matches/feedback": {
      "post": {
        "tags": ["matches"],
        "summary": "Submit match recommendation feedback",
        "description": "Stores a FeedbackRecord. The developer accepts (accepted=true) or rejects (accepted=false) a match recommendation. Rejection counts accumulate for GET /analytics/rejection-rate. The developer_id in the request body must match the authenticated user's JWT sub claim.",
        "operationId": "submitMatchFeedback",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FeedbackRequest" },
              "example": {
                "developer_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                "match_id": "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d",
                "accepted": false,
                "comment": "Project timezone does not work for me"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback stored",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FeedbackResponse" }
              }
            }
          },
          "400": {
            "description": "Malformed request body",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden — developer_id in body does not match authenticated user",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "match_id not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "422": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
    "/teams/{team_id}/risk-summary": {
      "get": {
        "tags": ["risk"],
        "summary": "Get risk badge summary for a team",
        "description": "Returns per-member risk badges (bench, burnout) and team-level distribution counts. Used by the Manager Dashboard. Raw behavioral vectors are never included in the response. Manager role required.",
        "operationId": "getTeamRiskSummary",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Team identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Team risk summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TeamRiskSummary" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden — manager role required",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Team not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/risk/refresh": {
      "post": {
        "tags": ["risk"],
        "summary": "Batch refresh all developer risk scores",
        "description": "Re-computes burnout_risk_score and bench_risk_score for all developers from current AllocationRecord data. Returns HTTP 202; processing is asynchronous. Manager role required.",
        "operationId": "refreshRiskScores",
        "responses": {
          "202": {
            "description": "Risk refresh job accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AsyncJobResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden — manager role required",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/admin/erasure-audit/{developer_id}": {
      "get": {
        "tags": ["admin"],
        "summary": "Get GDPR erasure audit log for a developer",
        "description": "Returns the ErasureAuditLog entry for the specified developer_id for compliance reporting. Records status 'completed' or 'failed'. Manager role required.",
        "operationId": "getErasureAudit",
        "parameters": [
          { "$ref": "#/components/parameters/DeveloperIdPath" }
        ],
        "responses": {
          "200": {
            "description": "Erasure audit record found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErasureAuditRecord" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden — manager role required",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "No erasure audit record found for this developer_id (erasure never requested)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/admin/reembed": {
      "post": {
        "tags": ["admin"],
        "summary": "Trigger full re-embedding of all profiles",
        "description": "Enqueues batch re-embedding for all DeveloperProfile and ProjectProfile records. Used when the embedding model changes. After completion, the HNSW index must be rebuilt offline (REINDEX CONCURRENTLY). Returns HTTP 202 Accepted. Manager role required.",
        "operationId": "triggerReembed",
        "responses": {
          "202": {
            "description": "Re-embedding job accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AsyncJobResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden — manager role required",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from POST /auth/login. Algorithm: HS256. Payload: {sub: user_id, role: 'developer'|'manager', iat, exp}. TTL: 1h. Secret loaded from env NEURAL_SYNC_JWT_SECRET. Per ADR-002: store in memory (not localStorage) on the React frontend; refresh token stored in HttpOnly cookie."
      }
    },
    "parameters": {
      "DeveloperIdPath": {
        "name": "developer_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "Developer unique identifier (UUID v4)"
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error envelope returned on all 4xx and 5xx responses.",
        "required": ["error_code", "message", "request_id"],
        "properties": {
          "error_code": {
            "type": "string",
            "description": "Machine-readable error code. Examples: NOT_FOUND, INVALID_WEIGHTS, VALIDATION_ERROR, ERASURE_FAILED, SERVICE_UNAVAILABLE, FORBIDDEN, UNAUTHORIZED, CONFLICT"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the error, safe to display in developer tooling"
          },
          "request_id": {
            "type": "string",
            "description": "Unique request trace identifier for correlating logs"
          }
        }
      },
...
      "TeamRiskMember": {
        "type": "object",
        "required": ["developer_id", "burnout_risk_badge", "bench_risk_badge", "burnout_risk_score", "bench_risk_score"],
        "properties": {
          "developer_id": { "type": "string", "format": "uuid" },
          "burnout_risk_score": { "type": "number", "minimum": 0.0, "maximum": 1.0 },
          "bench_risk_score": { "type": "number", "minimum": 0.0, "maximum": 1.0 },
          "burnout_risk_badge": { "type": "string", "enum": ["low", "medium", "high"] },
          "bench_risk_badge": { "type": "string", "enum": ["low", "medium", "high"] }
        }
      },
      "TeamRiskDistribution": {
        "type": "object",
        "properties": {
          "burnout_high_count": { "type": "integer", "minimum": 0 },
          "burnout_medium_count": { "type": "integer", "minimum": 0 },
          "burnout_low_count": { "type": "integer", "minimum": 0 },
          "bench_high_count": { "type": "integer", "minimum": 0 },
          "bench_medium_count": { "type": "integer", "minimum": 0 },
          "bench_low_count": { "type": "integer", "minimum": 0 }
        }
      },
      "TeamRiskSummary": {
        "type": "object",
        "description": "Aggregated risk view for Manager Dashboard. Raw behavioral vectors never included per requirements non-goal.",
        "required": ["team_id", "member_count", "members", "risk_distribution", "computed_at"],
        "properties": {
          "team_id": { "type": "string", "format": "uuid" },
          "member_count": { "type": "integer", "minimum": 0 },
          "members": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TeamRiskMember" }
          },
          "risk_distribution": { "$ref": "#/components/schemas/TeamRiskDistribution" },
          "computed_at": { "type": "string", "format": "date-time" }
        }
      },
      "FeedbackRequest": {
        "type": "object",
        "description": "Request body for POST /matches/feedback.",
        "required": ["developer_id", "match_id", "accepted"],
        "properties": {
          "developer_id": {
            "type": "string",
            "format": "uuid",
            "description": "Developer submitting feedback. Must match the authenticated JWT sub claim."
          },
          "match_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the match recommendation being responded to"
          },
          "accepted": {
            "type": "boolean",
            "description": "true = developer accepts the recommendation; false = developer rejects it. Rejections accumulate in rejection_ratio analytics."
          },
          "comment": {
            "type": ["string", "null"],
            "maxLength": 500,
            "description": "Optional rejection reason or note (free-text)"
          }
        }
      },
      "FeedbackResponse": {
        "type": "object",
        "required": ["id", "developer_id", "match_id", "accepted", "feedback_timestamp"],
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "FeedbackRecord UUID" },
          "developer_id": { "type": "string", "format": "uuid" },
          "match_id": { "type": "string", "format": "uuid" },
          "accepted": { "type": "boolean" },
          "comment": { "type": ["string", "null"] },
          "feedback_timestamp": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErasureAuditRecord": {
        "type": "object",
        "description": "ErasureAuditLog entry created on each DELETE /developers/{developer_id} request.",
        "required": ["erasure_request_id", "developer_id", "requested_at", "status", "initiating_user_id"],
        "properties": {
          "erasure_request_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for this erasure request"
          },
          "developer_id": { "type": "string", "format": "uuid" },
          "requested_at": { "type": "string", "format": "date-time" },
          "completed_at": {
            "type": ["string", "null"],
            "format": "date-time",
            "description": "Timestamp of successful cascade deletion; null if status=failed"
          },
          "status": {
            "type": "string",
            "enum": ["completed", "failed"],
            "description": "completed=all 6 cascade operations succeeded and committed; failed=transaction rolled back, developer data intact"
          },
          "initiating_user_id": {
            "type": "string",
            "format": "uuid",
            "description": "JWT sub of the user who initiated the erasure"
          }
        }
      },
...
      "AsyncJobResponse": {
        "type": "object",
        "description": "Generic HTTP 202 response for asynchronous background jobs.",
        "required": ["job_id", "message"],
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the queued background job"
          },
          "message": {
            "type": "string",
            "description": "Human-readable status message"
          },
          "estimated_count": {
            "type": ["integer", "null"],
            "description": "Estimated number of records to be processed (informational)"
          }
        }
      }
    }
  }
}